Rename to django-perf-rec
Save some typing.
