+++ README.rst
@@ -1,11 +1,11 @@
-===========================
-django-performance-recorder
-===========================
+===============
+django-perf-rec
+===============
 
-.. image:: https://img.shields.io/pypi/v/django-performance-recorder.svg
-    :target: https://pypi.python.org/pypi/django-performance-recorder
+.. image:: https://img.shields.io/pypi/v/django-perf-rec.svg
+    :target: https://pypi.python.org/pypi/django-perf-rec
 
-.. image:: https://img.shields.io/travis/YPlan/django-performance-recorder/master.svg
-        :target: https://travis-ci.org/YPlan/django-performance-recorder
+.. image:: https://img.shields.io/travis/YPlan/django-perf-rec/master.svg
+        :target: https://travis-ci.org/YPlan/django-perf-rec
 
 Keep detailed records of the performance of your Django code.
