+++ setup.cfg
@@ -10,13 +10,13 @@ add_imports =
     from __future__ import division
     from __future__ import print_function
     from __future__ import unicode_literals
-known_first_party = django_performance_recorder
+known_first_party = django_perf_rec
 known_third_party = django,yaml
 line_length = 120
 multi_line_output = 5
 not_skip = __init__.py
 
 [multilint]
-paths = django_performance_recorder
+paths = django_perf_rec
         setup.py
         tests
