+++ tests/test_sql.py
@@ -1,7 +1,7 @@
 # -*- coding:utf-8 -*-
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from django_performance_recorder.sql import sql_fingerprint
+from django_perf_rec.sql import sql_fingerprint
 
 
 def test_select():
